src/requirements.txt
setuptools~=56.0.0
psycopg2~=2.9.2
SQLAlchemy~=1.3.15
aiogram~=2.16
beautifulsoup4~=4.10.0
requests~=2.26.0
PyYAML~=6.0
typing~=3.7.4.3